Add Othello game logic, board representation, and visualization utilities - Begin implementing get_possible_moves - Restructure project for improved modularity - Enhance tests for OthelloRules to validate initial board setup and possible moves.

+++ include/othello/GameState.hpp
@@ -5,9 +5,10 @@
 #ifndef GAMESTATE_HPP
 #define GAMESTATE_HPP
 
-#include "GameBoard.hpp"
+#include "othello/GameBoard.hpp"
 #include <vector>
 #include <optional>
+#include <cstdint> // for int8_t
 
 namespace othello {
 
@@ -16,7 +17,7 @@ namespace othello {
   ///          whether the move is a pass. The position is represented as an
   ///         integer, where the board is indexed from 0 to 63 (for an 8x8 board).
   struct Move {
-    int position;  // the position of the move on the board
+    int8_t position;  // the position of the move on the board
     Color color;  // the color of the player making the move
     bool is_pass;  // flag to indicate if the move is a pass
   };
@@ -32,8 +33,8 @@ namespace othello {
     Color to_move;  // The color of the player whose turn it is to move
     std::optional<Color> winner = std::nullopt;  // The winner of the game (if any)
     bool game_over;  // Flag to indicate if the game is over
-    int black_score = 0;  // Score for black player
-    int white_score = 0;  // Score for white player
+    int8_t black_score = 0;  // Score for black player
+    int8_t white_score = 0;  // Score for white player
     void process_move(const Move &move);  // Process a move and update the game state accordingly
   };
